+++ client-core/src/test/java/com/influxdb/exceptions/InfluxExceptionTest.java
@@ -322,6 +322,32 @@ class InfluxExceptionTest {
                 .matches((Predicate<Throwable>) throwable -> throwable.toString().equals("com.influxdb.exceptions.InfluxException: HTTP status code: 501; Message: Wrong query"));
     }
 
+    @Test
+    void exceptionContainsHttpResponseHeaders() {
+      Assertions.assertThatThrownBy(() -> {
+        Response<Object> response = errorResponse(
+          "not found",
+          404,
+          15,
+          "not-json",
+          "X-Platform-Error-Code",
+          Map.of("Retry-After", "145",
+            "Trace-ID", "1234567989ABCDEF0",
+            "X-Influxdb-Build", "OSS"));
+        throw new InfluxException(new HttpException(response));
+        }
+      ).matches((Predicate<Throwable>) throwable -> ((InfluxException) throwable).status() == 404)
+       .matches((Predicate<Throwable>) throwable -> throwable.getMessage().equals(
+         "HTTP status code: 404; Message: not found"
+       ))
+       .matches((Predicate<Throwable>) throwable -> ((InfluxException) throwable).headers().size() == 5)
+       .matches((Predicate<Throwable>) throwable -> ((InfluxException) throwable).headers().get("Retry-After").equals("145"))
+       .matches((Predicate<Throwable>) throwable -> ((InfluxException) throwable).headers().get("X-Influxdb-Build").equals("OSS"))
+       .matches((Predicate<Throwable>) throwable -> ((InfluxException) throwable).headers().get("X-Influx-Reference").equals("15"))
+       .matches((Predicate<Throwable>) throwable -> ((InfluxException) throwable).headers().get("X-Platform-Error-Code").equals("not found"))
+       .matches((Predicate<Throwable>) throwable -> ((InfluxException) throwable).headers().get("Trace-ID").equals("1234567989ABCDEF0"));
+    }
+
     @Nonnull
     private Response<Object> errorResponse(@Nullable final String influxError) {
         return errorResponse(influxError, 500);